pyproject.toml
[tool.poetry]
name = "exam-scheduler-api"
version = "0.1.0"
description = ""
authors = ["zogon012 <zogon012@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
django-filter = "^24.2"
celery = "^5.4.0"
drf-yasg = "^1.21.7"
setuptools = "^70.1.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.4.4,<0.5.0"
coverage = "^7.5.1"
pytest = "^8.2.2"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"
pytest-cov = "^5.0.0"

[tool.ruff]
target-version = "py312"
line-length = 120
exclude = ["manage.py", "migrations", "templates", "dist", "build", "venv", "env", ".venv", ".env", ".git", ".pytest_cache", ".vscode", ".idea", "__pycache__"]

[tool.ruff.format]
quote-style = "single"

[tool.coverage.run]
omit = ["*/migrations/*"]
